Split out patch-specific functions.
